+++ packages/eyes-cypress/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+- updated to @applitools/dom-snapshot@4.3.1 (from 4.2.6)
+- updated to @applitools/eyes-sdk-core@12.9.1 (from 12.6.2)
+- updated to @applitools/visual-grid-client@15.3.0 (from 15.2.1)
 
 
 ## 3.16.1 - 2020/11/30